refac: on_event 데코레이터 제거 & Middleware 초기화 코드 이전
